Split qrcode.js into 3 files: qrHandler.js, utils.js, and web3Storage.js
